Fixing minor typo in AbstractServerPlugin update method should be int

+++ l2junity-gameserver/src/main/java/org/l2junity/gameserver/plugins/AbstractServerPlugin.java
@@ -152,7 +152,7 @@ public abstract class AbstractServerPlugin
 	 * @param fromVersion
 	 * @return
 	 */
-	public boolean update(double fromVersion)
+	public boolean update(int fromVersion)
 	{
 		return true;
 	}
